+++ static-assets/components/cstudio-forms/data-sources/file-browse-repo.js
@@ -80,29 +80,27 @@ YAHOO.extend(CStudioForms.Datasources.FileBrowseRepo, CStudioForms.CStudioFormDa
   },
 
   edit: function (key) {
-    var getContentItemCb = {
-      success: function (contentTO) {
-        var editCallback = {
-          success: function () {
-            // update label?
-          },
-          failure: function () {}
-        };
+    craftercms.services.content.fetchSandboxItem(CStudioAuthoringContext.site, key).subscribe({
+      next(sandboxItem) {
+        const readonly = !sandboxItem.availableActionsMap.edit;
 
-        CStudioAuthoring.Operations.editContent(
-          contentTO.item.contentType,
-          CStudioAuthoringContext.siteId,
-          contentTO.item.mimeType,
-          contentTO.item.nodeRef,
-          contentTO.item.uri,
-          false,
-          editCallback
-        );
+        if (readonly) {
+          CStudioAuthoring.Operations.showPreviewAsset(sandboxItem);
+        } else {
+          CStudioAuthoring.Operations.editContent(
+            sandboxItem.contentTypeId,
+            CStudioAuthoringContext.siteId,
+            sandboxItem.mimeType,
+            null,
+            sandboxItem.path,
+            false
+          );
+        }
       },
-      failure: function () {}
-    };
-
-    CStudioAuthoring.Service.lookupContentItem(CStudioAuthoringContext.site, key, getContentItemCb);
+      error(error) {
+        console.error(error?.response?.response);
+      }
+    });
   },
 
   getLabel: function () {