+++ ui/app/src/CHANGELOG.md
@@ -8,6 +8,8 @@
   * @mui/x-data-grid
   * @mui/x-date-pickers
   * @mui/x-tree-view
+* [utils/content]:
+  * Moved `PathNavigator/utils.ts` `isVideo`, `isAudio` and `isPdfDocument` functions to `utils/content.ts`.
 
 ## 4.2.0